Train/requirements.txt
torch
torchvision
albumentations
PILLOW
scikit-learn
tqdm
